requirements.txt
streamlit
streamlit-aggrid
streamlit-elements
psycopg2-binary
pytest
python-dotenv 